# backend/openapi/k12-api-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "K12 平台接口文档",
    "description": "K12 多智能体教学平台可调用接口。包含 IAM 账号安全、课程、智能体、结构化作业题目、答题与批改，统一通过 Gateway 和 JWT 鉴权。",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "网关服务"
    },
    {
      "url": "http://localhost:8081",
      "description": "用户与权限服务"
    },
    {
      "url": "http://localhost:8082",
      "description": "课程资源服务"
    },
    {
      "url": "http://localhost:8083",
      "description": "AI 智能体服务"
    },
    {
      "url": "http://localhost:8084",
      "description": "作业与批改服务"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/iam/auth/login": {
      "post": {
        "tags": [
          "01-用户与权限服务"
        ],
        "summary": "用户名密码登录并获取 JWT",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/iam/auth/register": {
      "post": {
        "tags": [
          "01-用户与权限服务"
        ],
        "summary": "注册学生账号",
        "operationId": "registerStudent",
        "description": "公开注册固定分配 ROLE_STUDENT，不能通过请求指定管理员角色。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
    "/api/v1/iam/roles": {
      "get": {
        "tags": [
          "01-用户与权限服务"
        ],
        "summary": "角色与权限列表",
        "operationId": "listRoles",
        "description": "需要 ROLE_ADMIN 或 role:read 权限。",
        "responses": {
          "200": {
            "description": "角色与权限列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListApiResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/iam/roles/{id}/permissions": {
      "put": {
        "tags": [
          "01-用户与权限服务"
        ],
        "summary": "更新角色权限",
        "operationId": "updateRolePermissions",
        "description": "需要 ROLE_ADMIN 或 role:update 权限。ROLE_ADMIN 权限不可通过该接口移除。",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolePermissionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的角色",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/iam/auth/token-state": {
      "get": {
        "tags": [
          "01-用户与权限服务"
        ],
        "summary": "校验当前 JWT 状态",
        "operationId": "validateTokenState",
        "description": "业务服务内部调用；JWT 的 authVersion 必须与数据库一致。",
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "example": 1
        }
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "未登录、JWT 无效或 JWT 已过期",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            },
            "example": {
              "code": 401,
              "message": "未登录、登录已过期或令牌无效",
              "data": null,
              "timestamp": "2026-08-14T02:30:00Z"
            }
          }
        }
      },
      "Forbidden": {
        "description": "已登录，但没有访问该资源的权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            },
            "example": {
              "code": 403,
              "message": "没有访问该资源的权限",
              "data": null,
              "timestamp": "2026-08-14T02:30:00Z"
            }
          }
        }
      },
      "BadRequest": {
        "description": "请求参数或业务输入不合法",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "唯一数据重复或资源状态冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiResponseBase": {
        "type": "object",
        "required": [
          "code",
          "message",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EmptyApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "nullable": true
              }
            }
          }
        ]
      },
...
      "UserUpdateRequest": {
        "type": "object",
        "required": [
          "username",
          "nickname",
          "roleCode"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "teacher01"
          },
          "nickname": {
            "type": "string",
            "example": "Teacher One"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "teacher01@example.com"
          },
          "roleCode": {
            "type": "string",
            "example": "ROLE_TEACHER"
          }
        }
      },
      "User": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserUpdateRequest"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64",
                "example": 1
              },
              "status": {
                "type": "string",
                "example": "ENABLED"
              },
              "updatedTime": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "UserApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        ]
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "admin123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT 访问令牌"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "example": 1800
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "authorities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "ROLE_ADMIN",
              "user:read",
              "user:create"
            ]
          }
        }
      },
      "LoginApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/LoginResponse"
              }
            }
          }
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "username",
          "password",
          "nickname"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "student01"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "maxLength": 72,
            "example": "Student@123"
          },
          "nickname": {
            "type": "string",
            "example": "学生一"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "student01@example.com"
          }
        }
      },
      "UpdateRolePermissionsRequest": {
        "type": "object",
        "required": [
          "permissionCodes"
        ],
        "properties": {
          "permissionCodes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "course:read",
              "course:update"
            ]
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 2
          },
          "name": {
            "type": "string",
            "example": "Teacher"
          },
          "code": {
            "type": "string",
            "example": "ROLE_TEACHER"
          },
          "description": {
            "type": "string",
            "example": "Teacher user"
          },
          "permissionCodes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "course:read",
              "course:update"
            ]
          }
        }
      },
      "RoleApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        ]
      },
      "RoleListApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        ]
      },
...
      "BooleanApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "boolean"
              }
            }
          }
        ]
      },
...
    }
  }
}